feat: ready to publish new `brainpylib` version
