+++ extensions/brainpylib/__init__.py
@@ -10,4 +10,4 @@ from .event_sum import *
 from .event_prod import *
 from .atomic_sum import *
 from .atomic_prod import *
-from custom_op import register_op
+from .custom_op import register_op
