+++ extensions/brainpylib/custom_op/__init__.py
@@ -1,3 +1,1 @@
-from cpu import *
-from gpu import *
-from regis_op import register_op
+from .regis_op import register_op
